Moved interface declarations to own file
