+++ frontend/src/Pages/Landing/Landing.tsx
@@ -1,25 +1,9 @@
 import { Box } from "@mui/material"
 import { useState, useEffect } from 'react';
-
 import CitySearchForm from './Forms/CitySearchForm';
 import FilterForm from './Forms/FilterForm';
 import MapContainer from "./MapContainer";
-
-export interface IPointsOfInterest {
-    title: string;
-    category: string;
-    description: string;
-    longitude: number;
-    latitude: number;
-    price: string;
-    city: string;
-    website: string;
-    post_code: string;
-    province: string;
-    country: string;
-    phone_number: number;
-    userId: number;
-}
+import { IPointsOfInterest } from "../../types/interfaces";
 
 const Landing = () => {
     // We might be able to move the city/setCity declaration into the CitySearchForm once we no longer use the StateChecker component in FilterForm, as that is the only other component that requires the city state
