+++ frontend/src/Pages/Landing/MapContainer.tsx
@@ -1,10 +1,10 @@
 import { useRef, useEffect, useState } from 'react';
 import mapboxgl from 'mapbox-gl';
 import { Box } from '@mui/material';
+import { IPointsOfInterest } from '../../types/interfaces';
 mapboxgl.accessToken = import.meta.env.VITE_MAPBOX_TOKEN;
 type MarkerWithId = mapboxgl.Marker & { id: string };
 type PopupWithId = mapboxgl.Popup & { id: string };
-import { IPointsOfInterest } from './Landing';
 
 type Props = {
     longitude: number;
@@ -35,7 +35,7 @@ const MapComponent = ({ longitude, latitude, renderedPointsOfInterest }: Props) 
             const marker = new mapboxgl.Marker() as MarkerWithId;
             marker.setLngLat([poi.longitude, poi.latitude]);
             marker.addTo(map.current);
-        };
+        }
 
     }, [longitude, latitude, renderedPointsOfInterest])
